+++ meteor/server/api/client.ts
@@ -64,7 +64,7 @@ export namespace ServerClientAPI {
 			userEvent,
 			eventTime,
 			`worker.${jobName}`,
-			[jobArguments],
+			jobArguments as any,
 			async (_credentials, userActionMetadata) => {
 				checkArgs()
 
@@ -91,7 +91,7 @@ export namespace ServerClientAPI {
 			userEvent,
 			eventTime,
 			`worker.${jobName}`,
-			[jobArguments],
+			jobArguments as any,
 			async (_credentials, userActionMetadata) => {
 				checkArgs()
 
@@ -111,7 +111,7 @@ export namespace ServerClientAPI {
 		playlistId: RundownPlaylistId,
 		checkArgs: () => void,
 		methodName: string,
-		args: any[],
+		args: Record<string, unknown>,
 		fcn: (access: VerifiedRundownPlaylistContentAccess) => Promise<T>
 	): Promise<ClientAPI.ClientResponse<T>> {
 		return runUserActionInLog(context, userEvent, eventTime, methodName, args, async () => {
@@ -132,7 +132,7 @@ export namespace ServerClientAPI {
 		rundownId: RundownId,
 		checkArgs: () => void,
 		methodName: string,
-		args: any[],
+		args: Record<string, unknown>,
 		fcn: (access: VerifiedRundownContentAccess) => Promise<T>
 	): Promise<ClientAPI.ClientResponse<T>> {
 		return runUserActionInLog(context, userEvent, eventTime, methodName, args, async () => {
@@ -184,7 +184,7 @@ export namespace ServerClientAPI {
 		userEvent: string,
 		eventTime: Time,
 		methodName: string,
-		methodArgs: unknown[],
+		methodArgs: Record<string, unknown>,
 		fcn: (credentials: BasicAccessContext, userActionMetadata: UserActionMetadata) => Promise<TRes>
 	): Promise<ClientAPI.ClientResponse<TRes>> {
 		// If we are in the test write auth check mode, then bypass all special logic to ensure errors dont get mangled