+++ meteor/server/api/rest/v1/studios.ts
@@ -176,7 +176,7 @@ class StudiosServerAPI implements StudiosRestAPI {
 			event,
 			getCurrentTime(),
 			'switchRouteSet',
-			[studioId, routeSetId, state],
+			{ studioId, routeSetId, state },
 			async () => {
 				check(studioId, String)
 				check(routeSetId, String)