+++ packages/job-worker/src/ingest/lock.ts
@@ -35,6 +35,7 @@ export interface CommitIngestData {
 }
 
 export enum UpdateIngestRundownAction {
+	REJECT = 'reject',
 	DELETE = 'delete',
 }
 
@@ -77,9 +78,9 @@ export async function runIngestJob(
 		const updatedIngestRundown = updateCacheFcn(clone(oldIngestRundown))
 		let newIngestRundown: LocalIngestRundown | undefined
 		switch (updatedIngestRundown) {
-			// case UpdateIngestRundownAction.REJECT:
-			// 	// Reject change
-			// 	return
+			case UpdateIngestRundownAction.REJECT:
+				// Reject change
+				return rundownId
 			case UpdateIngestRundownAction.DELETE:
 				ingestObjCache.delete()
 				newIngestRundown = undefined