+++ packages/job-worker/src/playout/__tests__/timeline.test.ts
@@ -1240,7 +1240,7 @@ describe('Timeline', () => {
 						})
 					)
 
-					const adlibbedPieceId = 'randomId9010'
+					const adlibbedPieceId = 'randomId9007'
 
 					// The adlib should be starting at 'now'
 					await checkTimings({
@@ -1406,7 +1406,7 @@ describe('Timeline', () => {
 						})
 					)
 
-					const adlibbedPieceId = 'randomId9010'
+					const adlibbedPieceId = 'randomId9007'
 
 					// The adlib should be starting at 'now'
 					await checkTimings({