+++ packages/job-worker/src/playout/adlibAction.ts
@@ -195,10 +195,11 @@ export async function executeActionInner(
 	// If any action cannot be done due to timings, that needs to be rejected by the context
 	if (!blueprint.blueprint.executeAction) throw UserError.create(UserErrorMessage.ActionsNotSupported)
 
-	logger.info(
-		`Executing AdlibAction "${actionParameters.actionId}": ${JSON.stringify(actionParameters.userData)} (${
-			actionParameters.triggerMode
-		})`
+	logger.info(`Executing AdlibAction "${actionParameters.actionId}"`)
+	logger.silly(
+		`Executing AdlibAction Payload "${actionParameters.actionId}" Payload: ${JSON.stringify(
+			actionParameters.userData
+		)} (${actionParameters.triggerMode})`
 	)
 
 	try {
@@ -264,11 +265,9 @@ async function executeDataStoreAction(
 			showStyle,
 			watchedPackages
 		)
-
-		logger.info(
-			`Executing Datastore AdlibAction "${actionParameters.actionId}": ${JSON.stringify(
-				actionParameters.userData
-			)}`
+		logger.info(`Executing Datastore AdlibAction "${actionParameters.actionId}"`)
+		logger.silly(
+			`Datastore AdlibAction "${actionParameters.actionId}" Payload: ${JSON.stringify(actionParameters.userData)}`
 		)
 
 		try {