+++ packages/job-worker/src/studio/model/StudioPlayoutModelImpl.ts
@@ -84,7 +84,10 @@ export class StudioPlayoutModelImpl implements StudioPlayoutModel {
 		this.#baselineHelper.setExpectedPlayoutItems(playoutItems)
 	}
 
-	setTimeline(timelineObjs: TimelineObjGeneric[], generationVersions: TimelineCompleteGenerationVersions): void {
+	setTimeline(
+		timelineObjs: TimelineObjGeneric[],
+		generationVersions: TimelineCompleteGenerationVersions
+	): ReadonlyDeep<TimelineComplete> {
 		this.#timeline = {
 			_id: this.context.studioId,
 			timelineHash: getRandomId(),
@@ -93,6 +96,8 @@ export class StudioPlayoutModelImpl implements StudioPlayoutModel {
 			generationVersions: generationVersions,
 		}
 		this.#timelineHasChanged = true
+
+		return this.#timeline
 	}
 
 	/**