+++ packages/webui/src/client/ui/Prompter/prompter.ts
@@ -1,13 +1,4 @@
-import * as _ from 'underscore'
 import { ScriptContent, SourceLayerType } from '@sofie-automation/blueprints-integration'
-import { normalizeArrayToMap, protectString } from '../../lib/tempLib'
-import { Piece } from '@sofie-automation/corelib/dist/dataModel/Piece'
-import { getPieceInstancesForPartInstance, getSegmentsWithPartInstances } from '../../lib/RundownResolver'
-import { FindOptions } from '../../collections/lib'
-import { PieceInstance } from '@sofie-automation/corelib/dist/dataModel/PieceInstance'
-import { Rundown } from '@sofie-automation/corelib/dist/dataModel/Rundown'
-import { processAndPrunePieceInstanceTimings } from '@sofie-automation/corelib/dist/playout/processAndPrune'
-import { UIShowStyleBases } from '../Collections'
 import {
 	PartId,
 	PartInstanceId,
@@ -17,10 +8,19 @@ import {
 	SegmentId,
 	ShowStyleBaseId,
 } from '@sofie-automation/corelib/dist/dataModel/Ids'
-import { RundownPlaylists, PieceInstances, Pieces, Segments } from '../../collections'
-import { SourceLayers } from '@sofie-automation/corelib/dist/dataModel/ShowStyleBase'
-import { RundownPlaylistCollectionUtil } from '../../collections/rundownPlaylistUtil'
+import { Piece } from '@sofie-automation/corelib/dist/dataModel/Piece'
+import { PieceInstance } from '@sofie-automation/corelib/dist/dataModel/PieceInstance'
+import { Rundown } from '@sofie-automation/corelib/dist/dataModel/Rundown'
 import { DBSegment } from '@sofie-automation/corelib/dist/dataModel/Segment'
+import { SourceLayers } from '@sofie-automation/corelib/dist/dataModel/ShowStyleBase'
+import { processAndPrunePieceInstanceTimings } from '@sofie-automation/corelib/dist/playout/processAndPrune'
+import * as _ from 'underscore'
+import { FindOptions } from '../../collections/lib'
+import { RundownPlaylistCollectionUtil } from '../../collections/rundownPlaylistUtil'
+import { normalizeArrayToMap, protectString } from '../../lib/tempLib'
+import { PieceInstances, Pieces, RundownPlaylists, Segments } from '../../collections'
+import { getPieceInstancesForPartInstance, getSegmentsWithPartInstances } from '../../lib/RundownResolver'
+import { UIShowStyleBases } from '../Collections'
 
 // export interface NewPrompterAPI {
 // 	getPrompterData (playlistId: RundownPlaylistId): Promise<PrompterData>
@@ -29,9 +29,15 @@ import { DBSegment } from '@sofie-automation/corelib/dist/dataModel/Segment'
 // 	'getPrompterData' = 'PrompterMethods.getPrompterData'
 // }
 
+export interface PrompterDataRundown {
+	id: RundownId
+	title: string
+	segments: PrompterDataSegment[]
+}
+
 export interface PrompterDataSegment {
 	id: SegmentId
-	title: string | undefined
+	title: string
 	parts: PrompterDataPart[]
 }
 export interface PrompterDataPart {
@@ -47,7 +53,7 @@ export interface PrompterData {
 	title: string
 	currentPartInstanceId: PartInstanceId | null
 	nextPartInstanceId: PartInstanceId | null
-	segments: Array<PrompterDataSegment>
+	rundowns: Array<PrompterDataRundown>
 }
 
 export namespace PrompterAPI {
@@ -112,7 +118,7 @@ export namespace PrompterAPI {
 			title: playlist.name,
 			currentPartInstanceId: currentPartInstance ? currentPartInstance._id : null,
 			nextPartInstanceId: nextPartInstance ? nextPartInstance._id : null,
-			segments: [],
+			rundowns: [],
 		}
 
 		const piecesIncluded: PieceId[] = []
@@ -169,6 +175,8 @@ export namespace PrompterAPI {
 			  ).fetch()
 			: undefined
 
+		const orderedRundowns = new Map<RundownId, PrompterDataRundown>()
+
 		groupedParts.forEach(({ segment, partInstances }, segmentIndex) => {
 			const segmentId = segment._id
 			const rundown = rundownMap.get(segment.rundownId)
@@ -181,9 +189,19 @@ export namespace PrompterAPI {
 				return
 			}
 
+			let rundownObj = orderedRundowns.get(rundown._id)
+			if (!rundownObj) {
+				rundownObj = {
+					id: rundown._id,
+					title: rundown.name,
+					segments: [],
+				}
+				orderedRundowns.set(rundown._id, rundownObj)
+			}
+
 			const segmentData: PrompterDataSegment = {
 				id: segmentId,
-				title: segment ? segment.name : undefined,
+				title: segment.name,
 				parts: [],
 			}
 
@@ -229,19 +247,17 @@ export namespace PrompterAPI {
 						const piece = pieceInstance.piece
 						const sourceLayer = sourceLayers[piece.sourceLayerId]
 
-						if (piece.content && sourceLayer && sourceLayer.type === SourceLayerType.SCRIPT) {
-							const content = piece.content as ScriptContent
-							if (content.fullScript) {
-								if (piecesIncluded.indexOf(piece._id) >= 0) {
-									break // piece already included in prompter script
-								}
-								piecesIncluded.push(piece._id)
-								partData.pieces.push({
-									id: piece._id,
-									text: content.fullScript,
-								})
-							}
-						}
+						if (!piece.content || !sourceLayer || sourceLayer.type !== SourceLayerType.SCRIPT) break
+
+						const content = piece.content as ScriptContent
+						if (!content.fullScript) break
+						if (piecesIncluded.indexOf(piece._id) >= 0) break // piece already included in prompter script
+
+						piecesIncluded.push(piece._id)
+						partData.pieces.push({
+							id: piece._id,
+							text: content.fullScript,
+						})
 					}
 				}
 
@@ -256,8 +272,11 @@ export namespace PrompterAPI {
 				segmentData.parts.push(partData)
 			}
 
-			data.segments.push(segmentData)
+			rundownObj.segments.push(segmentData)
 		})
+
+		data.rundowns = Array.from(orderedRundowns.values())
+
 		return data
 	}
 }